+++ ydb/apps/pgwire/pgwire.cpp
@@ -9,7 +9,7 @@
 #include <ydb/library/actors/core/scheduler_basic.h>
 #include <ydb/library/actors/core/process_stats.h>
 #include <ydb/library/actors/protos/services_common.pb.h>
-#include <ydb/library/actors/interconnect/poller_actor.h>
+#include <ydb/library/actors/interconnect/poller/poller_actor.h>
 #include "pgwire.h"
 #include "log_impl.h"
 #include "pg_ydb_proxy.h"
