+++ ydb/core/base/appdata.cpp
@@ -34,7 +34,7 @@
 #include <ydb/core/protos/workload_manager_config.pb.h>
 #include <ydb/library/pdisk_io/aio.h>
 
-#include <ydb/library/actors/interconnect/poller_tcp.h>
+#include <ydb/library/actors/interconnect/poller/poller_tcp.h>
 #include <ydb/library/actors/core/monotonic_provider.h>
 #include <ydb/library/actors/util/should_continue.h>
 #include <library/cpp/random_provider/random_provider.h>
