+++ ydb/core/base/appdata.h
@@ -35,7 +35,7 @@
 #include <ydb/core/base/event_filter.h>
 #include <ydb/library/actors/core/actor.h>
 
-#include <ydb/library/actors/interconnect/poller_tcp.h>
+#include <ydb/library/actors/interconnect/poller/poller_tcp.h>
 #include <ydb/library/actors/core/monotonic_provider.h>
 #include <ydb/library/actors/util/should_continue.h>
 #include <library/cpp/random_provider/random_provider.h>
