+++ ydb/core/driver_lib/cli_utils/cli.h
@@ -5,7 +5,7 @@
 #include <ydb/core/driver_lib/cli_base/cli.h>
 #include <ydb/core/driver_lib/run/factories.h>
 
-#include <ydb/library/actors/interconnect/poller_tcp.h>
+#include <ydb/library/actors/interconnect/poller/poller_tcp.h>
 #include <ydb/public/lib/deprecated/client/msgbus_client.h>
 
 #include <library/cpp/deprecated/enum_codegen/enum_codegen.h>
