+++ ydb/core/kafka_proxy/actors/actors.h
@@ -30,7 +30,7 @@ enum class EBalancingMode {
 struct TReadSession {
     EBalancingMode BalancingMode = EBalancingMode::Server;
     std::optional<EBalancingMode> PendingBalancingMode;
-    TActorId ProxyActorId;
+    NActors::TActorId ProxyActorId;
 };
 
 struct TContext {
@@ -47,7 +47,7 @@ struct TContext {
 
     const NKikimrConfig::TKafkaProxyConfig& Config;
 
-    TActorId ConnectionId;
+    NActors::TActorId ConnectionId;
     TString KafkaClient;
 
 
@@ -195,7 +195,7 @@ NActors::IActor* CreateKafkaApiVersionsActor(const TContext::TPtr context, const
 NActors::IActor* CreateKafkaInitProducerIdActor(const TContext::TPtr context, const ui64 correlationId, const TMessagePtr<TInitProducerIdRequestData>& message);
 NActors::IActor* CreateKafkaMetadataActor(const TContext::TPtr context, const ui64 correlationId,
                                           const TMessagePtr<TMetadataRequestData>& message,
-                                          const TActorId& discoveryCacheActor);
+                                          const NActors::TActorId& discoveryCacheActor);
 NActors::IActor* CreateKafkaProduceActor(const TContext::TPtr context);
 NActors::IActor* CreateKafkaReadSessionProxyActor(const TContext::TPtr context, ui64 cookie);
 NActors::IActor* CreateKafkaReadSessionActor(const TContext::TPtr context, ui64 cookie);
