+++ ydb/core/pgproxy/pg_listener.h
@@ -11,7 +11,7 @@ inline NActors::IActor* CreatePGListener(const NActors::TActorId& poller, const 
                                          const TListenerSettings& settings = {.Port = 5432}) {
     return CreateSocketListener(
         poller, settings,
-        [=](const TActorId& listenerActorId, TIntrusivePtr<TSocketDescriptor> socket, TNetworkConfig::TSocketAddressType address) {
+        [=](const NActors::TActorId& listenerActorId, TIntrusivePtr<TSocketDescriptor> socket, TNetworkConfig::TSocketAddressType address) {
             return CreatePGConnection(listenerActorId, socket, address, databaseProxy);
         },
         NKikimrServices::EServiceKikimr::PGWIRE);
