+++ ydb/core/raw_socket/sock_listener.cpp
@@ -1,7 +1,7 @@
 
 #include <ydb/library/actors/core/actor_bootstrapped.h>
 #include <ydb/library/actors/core/log.h>
-#include <ydb/library/actors/interconnect/poller_actor.h>
+#include <ydb/library/actors/interconnect/poller/poller_actor.h>
 #include <util/network/sock.h>
 #include <ydb/library/services/services.pb.h>
