+++ ydb/core/test_tablet/defs.h
@@ -8,6 +8,6 @@
 #include <ydb/core/protos/counters_testshard.pb.h>
 #include <ydb/core/protos/test_shard.pb.h>
 #include <ydb/library/actors/core/actor_coroutine.h>
-#include <ydb/library/actors/interconnect/poller_actor.h>
+#include <ydb/library/actors/interconnect/poller/poller_actor.h>
 #include <library/cpp/json/json_writer.h>
 #include <contrib/libs/t1ha/t1ha.h>
