+++ ydb/library/actors/core/ut/scheduler_actor_ut.cpp
@@ -6,7 +6,7 @@
 #include "events.h"
 #include "event_local.h"
 #include "hfunc.h"
-#include <ydb/library/actors/interconnect/poller_actor.h>
+#include <ydb/library/actors/interconnect/poller/poller_actor.h>
 #include <library/cpp/testing/unittest/registar.h>
 
 #include <util/system/sanitizers.h>
