+++ ydb/library/actors/interconnect/events_local.h
@@ -1,73 +1,15 @@
 #pragma once
 
-#include <ydb/library/actors/core/events.h>
 #include <ydb/library/actors/core/event_local.h>
 #include <ydb/library/actors/protos/interconnect.pb.h>
 #include <util/generic/deque.h>
 #include <util/network/address.h>
 
+#include "events/events.h"
 #include "interconnect_stream.h"
 #include "types.h"
 
 namespace NActors {
-    enum class ENetwork : ui32 {
-        ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        // local messages
-        ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-        Start = EventSpaceBegin(TEvents::ES_INTERCONNECT_TCP),
-
-        SocketReadyRead = Start,
-        SocketReadyWrite,
-        SocketError,
-        Connect,
-        Disconnect,
-        IncomingConnection,
-        HandshakeAsk,
-        HandshakeAck,
-        HandshakeNak,
-        HandshakeDone,
-        HandshakeFail,
-        Kick,
-        Flush,
-        NodeInfo,
-        BunchOfEventsToDestroy,
-        HandshakeRequest,
-        HandshakeReplyOK,
-        HandshakeReplyError,
-        ResolveAddress,
-        AddressInfo,
-        ResolveError,
-        HTTPStreamStatus,
-        HTTPSendContent,
-        ConnectProtocolWakeup,
-        HTTPProtocolRetry,
-        EvPollerRegister,
-        EvPollerRegisterResult,
-        EvPollerReady,
-        EvUpdateFromInputSession,
-        EvConfirmUpdate,
-        EvSessionBufferSizeRequest,
-        EvSessionBufferSizeResponse,
-        EvProcessPingRequest,
-        EvGetSecureSocket,
-        EvSecureSocket,
-        HandshakeBrokerTake,
-        HandshakeBrokerFree,
-        HandshakeBrokerPermit,
-
-        // external data channel messages
-        EvSubscribeForConnection,
-        EvReportConnection,
-
-        ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        // nonlocal messages; their indices must be preserved in order to work properly while doing rolling update
-        ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-        // interconnect load test message
-        EvLoadMessage = Start + 256,
-    };
-
     struct TEvSocketReadyRead: public TEventLocal<TEvSocketReadyRead, ui32(ENetwork::SocketReadyRead)> {
     };
