+++ ydb/library/actors/interconnect/interconnect_stream.cpp
@@ -1,11 +1,12 @@
 #include "interconnect_stream.h"
 #include "interconnect_common.h"
 #include "logging.h"
-#include "poller_actor.h"
 #include <util/network/socket.h>
 #include <openssl/ssl.h>
 #include <openssl/err.h>
 #include <openssl/pem.h>
+
+#include <ydb/library/actors/interconnect/poller/poller_actor.h>
 
 #if defined(_win_)
 #include <util/system/file.h>
