+++ ydb/library/actors/interconnect/ut/poller_actor_ut.cpp
@@ -1,4 +1,4 @@
-#include <ydb/library/actors/interconnect/poller_actor.h>
+#include <ydb/library/actors/interconnect/poller/poller_actor.h>
 #include <ydb/library/actors/testlib/test_runtime.h>
 
 #include <library/cpp/testing/unittest/registar.h>
