+++ ydb/library/actors/interconnect/ya.make
@@ -49,15 +49,6 @@ SRCS(
     logging.h
     packet.cpp
     packet.h
-    poller_actor.cpp
-    poller_actor.h
-    poller.h
-    poller_tcp.cpp
-    poller_tcp.h
-    poller_tcp_unit.cpp
-    poller_tcp_unit.h
-    poller_tcp_unit_select.cpp
-    poller_tcp_unit_select.h
     profiler.h
     slowpoke_actor.h
     subscription_manager.cpp
@@ -67,13 +58,6 @@ SRCS(
     watchdog_timer.h
 )
 
-IF (OS_LINUX)
-    SRCS(
-        poller_tcp_unit_epoll.cpp
-        poller_tcp_unit_epoll.h
-    )
-ENDIF()
-
 PEERDIR(
     contrib/libs/libc_compat
     contrib/libs/openssl
@@ -82,6 +66,7 @@ PEERDIR(
     ydb/library/actors/dnscachelib
     ydb/library/actors/dnsresolver
     ydb/library/actors/helpers
+    ydb/library/actors/interconnect/poller
     ydb/library/actors/prof
     ydb/library/actors/protos
     ydb/library/actors/util
