+++ ydb/library/yql/providers/dq/service/interconnect_helpers.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <ydb/library/actors/core/actorsystem.h>
-#include <ydb/library/actors/interconnect/poller_tcp.h>
+#include <ydb/library/actors/interconnect/poller/poller_tcp.h>
 #include <ydb/library/actors/interconnect/interconnect.h>
 #include <library/cpp/yson/node/node.h>
